Reorganized and renamed files related to Commission report
 - Moved all files to reports/commission folder
 - corrected spelling of commission.js
